리뷰페이지 - 별점 구현 완료 남은 것 목록화 끝마무리, 수정, 삭제 채팅 여부에 따른 리뷰 가능여부

+++ src/main/java/kbg/modu/moduproject/domain/Review.java
@@ -22,7 +22,7 @@ public class Review {
     //내용
     private String content;
     //별점
-    private Integer starScore;
+    private float starScore;
     //수정여부
     private String updateYn;
     //삭제여부
